+++ src/CognitiveServices/CognitiveServices/CognitiveServicesAccount/SetAzureCognitiveServicesAccount.cs
@@ -148,7 +148,7 @@ namespace Microsoft.Azure.Commands.Management.CognitiveServices
         [Parameter(
             Mandatory = false,
             HelpMessage = "The ApiProperties of Cognitive Services Account. Required by specific account types.")]
-        public CognitiveServicesAccountApiProperties ApiProperties { get; set; }
+        public CognitiveServicesAccountApiProperties ApiProperty { get; set; }
 
         [Parameter(Mandatory = false, HelpMessage = "Don't ask for confirmation.")]
         public SwitchParameter Force { get; set; }
@@ -169,10 +169,10 @@ namespace Microsoft.Azure.Commands.Management.CognitiveServices
                 hasPropertiesChange = true;
                 properties.NetworkAcls = NetworkRuleSet.ToNetworkRuleSet();
             }
-            if (ApiProperties != null)
+            if (ApiProperty != null)
             {
                 hasPropertiesChange = true;
-                properties.ApiProperties = ApiProperties;
+                properties.ApiProperties = ApiProperty;
             }
 
             Sku sku = null;